RMI fix

+++ AuthMicroservicio/src/main/java/com/FileDepot/Interfaces/AuthInterface.java
@@ -4,6 +4,7 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface AuthInterface extends Remote {
-boolean login(String email, String password) throws RemoteException;
-boolean register(String name, String email, String password, int number) throws RemoteException; 
+    boolean login(String email, String password) throws RemoteException;
+    boolean register(String name, String email, String password, long number) throws RemoteException;
+
 }
